+++ src/ResourceManager/Common/Commands.Common.Authentication.ResourceManager/Models/PSAzureTenant.cs
@@ -85,7 +85,7 @@ namespace Microsoft.Azure.Commands.Profile.Models
             this.PopulateExtensions(other);
         }
         /// <summary>
-        /// The subscription id.
+        /// The tenant id.
         /// </summary>
         public string Id { get; set; }
 
@@ -93,6 +93,11 @@ namespace Microsoft.Azure.Commands.Profile.Models
         /// The name of the subscription.
         /// </summary>
         public string Directory { get; set; }
+
+        /// <summary>
+        /// For legacy support - the tenant id
+        /// </summary>
+        public string TenantId { get { return Id; } }
 
         /// <summary>
         /// Extended proeprties of the tenant