+++ src/ResourceManager/Profile/Commands.Profile/Errors/AzureExceptionRecord.cs
@@ -23,9 +23,9 @@ namespace Microsoft.Azure.Commands.Profile.Errors
         {
             if (exception != null)
             {
-                Message = exception.Message;
-                HelpLink = exception.HelpLink;
-                StackTrace = exception.StackTrace;
+            	Message = exception.Message;
+            	HelpLink = exception.HelpLink;
+            	StackTrace = exception.StackTrace;
             }
 
             Exception = exception;