+++ src/ResourceManager/Relay/Commands.Relay/Models/RelayNamespaceAttributes.cs
@@ -12,6 +12,7 @@ namespace Microsoft.Azure.Commands.Relay.Models
     using Newtonsoft.Json;
     using System.Collections.Generic;
     using Azure.Management.Relay.Models;
+    using System.Text.RegularExpressions;
 
     /// <summary>
     /// Description of a Namespace resource.
@@ -58,12 +59,19 @@ namespace Microsoft.Azure.Commands.Relay.Models
             Name = relaynamespace.Name;
             Type = relaynamespace.Type;
             Tags = relaynamespace.Tags;
-            ProvisioningState = relaynamespace.ProvisioningState;
+            ProvisioningState = relaynamespace.ProvisioningState.ToString();
             CreatedAt = relaynamespace.CreatedAt;
             UpdatedAt = relaynamespace.UpdatedAt;
             ServiceBusEndpoint = relaynamespace.ServiceBusEndpoint;
             MetricId = relaynamespace.MetricId;
+            ResourceGroup = Regex.Split(relaynamespace.Id, @"/")[4];
         }
+
+
+        /// <summary>
+        /// Gets the resourcegroup name
+        /// </summary>
+        public string ResourceGroup { get; }
 
         /// <summary>
         /// Static constructor for RelayNamespaceAttributes class.