+++ src/ResourceManager/Websites/Commands.Websites/Cmdlets/WebApps/NewAzureWebApp.cs
@@ -12,11 +12,12 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-
-using Microsoft.Azure.Commands.Resources.Models;
 using Microsoft.Azure.Commands.WebApps.Models;
 using Microsoft.Azure.Commands.WebApps.Models.WebApp;
 using Microsoft.Azure.Commands.WebApps.Utilities;
+using Microsoft.Azure.Management.Internal.Resources;
+using Microsoft.Azure.Management.Internal.Resources.Models;
+using Microsoft.Azure.Management.Internal.Resources.Utilities;
 using Microsoft.Azure.Management.WebSites.Models;
 using Microsoft.Azure.Management.WebSites;
 using Microsoft.WindowsAzure.Commands.Common;
@@ -25,15 +26,6 @@ using System.Collections;
 using System.Linq;
 using System.Management.Automation;
 
-#if !NETSTANDARD
-using Microsoft.Azure.Management.Resources;
-using Microsoft.Azure.Management.Resources.Models;
-using ResourceManagerDeployment = Microsoft.Azure.Management.Resources.Models.Deployment;
-#else
-using Microsoft.Azure.Management.ResourceManager;
-using Microsoft.Azure.Management.ResourceManager.Models;
-using ResourceManagerDeployment = Microsoft.Azure.Management.ResourceManager.Models.Deployment;
-#endif
 namespace Microsoft.Azure.Commands.WebApps.Cmdlets.WebApps
 {
     /// <summary>
@@ -149,7 +141,7 @@ namespace Microsoft.Azure.Commands.WebApps.Cmdlets.WebApps
             var template = DeploymentTemplateHelper.CreateSlotCloneDeploymentTemplate(Location, AppServicePlan, Name, SourceWebApp.Id,
                 slotNames, hostingEnvironmentProfile, WebsitesClient.WrappedWebsitesClient.ApiVersion());
 
-            var deployment = new ResourceManagerDeployment
+            var deployment = new Management.Internal.Resources.Models.Deployment
             {
                 Properties = new DeploymentProperties
                 {