+++ src/ServiceManagement/Compute/Commands.ServiceManagement/IaaS/Extensions/DSC/DscExtensionCmdletCommonBase.cs
@@ -36,13 +36,13 @@ namespace Microsoft.WindowsAzure.Commands.ServiceManagement.IaaS.Extensions.DSC
         /// Attempts to get the user's credentials from the given Storage Context or the current subscription, if the former is null. 
         /// Throws a terminating error if the credentials cannot be determined.
         /// </summary>
-        internal static StorageCredentials GetStorageCredentials(this AzureSMCmdlet cmdlet, AzureStorageContext storageContext)
+        internal static StorageCredentials GetStorageCredentials(this AzureSMCmdlet cmdlet, IStorageContext storageContext)
         {
             StorageCredentials credentials = null;
 
             if (storageContext != null)
             {
-                credentials = storageContext.StorageAccount.Credentials;
+                credentials = storageContext.GetCloudStorageAccount().Credentials;
             }
             else
             {