+++ src/ResourceManager/AnalysisServices/ChangeLog.md
@@ -19,6 +19,17 @@
 -->
 ## Current Release
 
+## Version 0.4.4
+
+## Version 0.4.3
+* Fixed bug in Set-AzureRmAnalysisServciesServer
+    - When admin was not provided, the admin will be removed.
+* Added BackupBlobContainerUri in New-AzureRmAnalysisServicesServer and Set-AzureRmAnalysisServicesServer
+    - Enable to set/disable backup blob container for backup/restore Azure Analysis Services Server
+* Updated Sku lookup in New-AzureRmAnalysisServicesServer and Set-AzureRmAnalysisServicesServer
+    - Changed hard coded Sku into dynamic lookup.
+* Add-AzureAnalysisServicesAccount to support login with Service Principal
+
 ## Version 0.4.2
 
 ## Version 0.4.1