+++ src/ResourceManager/IotHub/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 ## Current Release
 
+## Version 2.3.1
+
+## Version 2.3.0
+
 ## Version 2.2.1
 
 ## Version 2.2.0